renamed the cookie file
